feat(overview): restore ext/downloads charts in v2 header; StyledLink stats
